Increased MitigationName limit to 200
- In order to support Vanta mitigationNames, increased the maximum length of mitigationName to be 200.

This value is introduced based on the below calculation of the mitigationName length of Vanta shared-objects:
1. Policy MitigationName example: P1000-firewall_123456789120_5ea1667b-9771-47f6-808e-eeb16506fa9e = 64 characters
2. RuleGroup MitigationName example: P2000-firewall_123456789120_5ea1667b-9771-47f6-808e-eeb16506fa9e_stateful-rulegroup_123456789120_7288fccc-2d16-468d-8eb9-aee846f6a965 = 134 characters
3. FirewallMetadata MitigationName example: P3000-firewall_123456789012_5ea1667b-9771-47f6-808e-eeb16506fa9e_meta = 69 characters

Based on the above calculation, the maximum length of mitigation name = 134, hence increasing the Max length of MitigationName to 200.
With the new naming scheme, the maximum length of mitigationFileName = 201 < 255.

The above naming scheme is described in : https://sim.amazon.com/issues/vanta-857

cr https://code.amazon.com/reviews/CR-13873145