Expose a strongly-consistent GetRequestStatus operation
Use the optional mitigationName request parameter to do a strongly
consistent read operation on the archive.  If the mitigation name
parameter is not provided, we will continue doing an eventually
consistent read operation against the index instead.

This change maintains backward compatibility with all existing callers
(which don't provide mitigation name) but also enables callers that do
know the mitigation name to do a strongly-consistent operation.  This is
important because due to the eventual consistency of the GSI, requests
can appear to momentarily disappear and be not found for a short time as
they are moved to the archive.

cr: https://code.amazon.com/reviews/CR-82457968